third_party/blink/web_tests/virtual/threaded-prefer-compositing/fast/scroll-behavior/smooth-scroll/fixed-background-in-iframe-expected.txt

This test verifies that smooth scrolls run on the main thread when there are fixed backgrounds on the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS animatedOnMainThread is true
PASS successfullyParsed is true

TEST COMPLETE
